feat: Added logging integration test
First logging integration test is passing with a simple log. Python script currently defaults to Android as only Android emulators work on the m1, iOS requires a physical device.

Next steps are to add parameters to the run_integration_test to select platforms and make sure the result.xml are in a better location.  Then we can proceed to make more thorough integration tests.

refs: RUMM-3275 RUMM-3274