feat: Use a background worker thread to dispatch logs
Because of the latency of calls both on JNI and on iOS, we've pushed the work of dispatching calls to native SDKs to a background worker thread. This significantly improves timing even in development builds.

I've used a proxy logger, so the platform DDLogers didn't need to change, and we can revert back to direct logging if needed.

refs: RUMM-3480