feat: Copy Symbols option for iOS
Add the "Copy Symbols" option to the Datadog Options.

When "Copy Symbols" is enabled, the build process will create a `datadogSymbols` directory and output a build_id file to it that identifies the build. It also enables `--emit-mapping-file` for IL2CPP and copies the resulting `LineNumberMapping.json` to `datadogSymbols`.

This also adds the iOS specific portions. when "Copy Symbols" is enabled, the build id is added as an additional configuration property, and a shell script is added to the XCode build to copy built dSYMS to `datadogSymbols`.

refs: RUM-2380