fix: Isolate Datadog to the `UnityFramework`
The Datadog SDK added itself to both the `Unity-iPhone` and `UnityFramework` targets in Cocoapods, because the Bridge needed to be in UnityFramework, and initialization was in `Unity-iPhone`. With static linking (EDM4U's default) this caused issues with two copies of the SDK being loaded at the same time.

Using Dynamic Linking caused some weird issues where PLCrashReporter couldn't be found when certain linker flags were set, and breaks some pods that require static linking.

So, we've isolated the DatadogSDK to UnityFramework by initializing it in `applicaiton:didFinishLaunchingWithOptions` instead of in `main` which allows us to use static linking (the EDM4U default) and not have duplicate SDK issues.