feat: Warn and discard null values passed to worker proxies
Certain values cannot or should not be null when passed to the Worker Proxy classes. While many of the system libraries check this, the logs they'll generate are too far from the call site to be useful. Move the checks further up into the worker proxies both for safety and also to show the warnings as close to the call site as possible.

refs: RUM-7947