feat: Support Logs in WebGL targets.
Adds the JS interop to support calling the Browser SDK from Unity. This relies heavily on JSON serialization at the moment to pass objects between Unity and the Browser SDK, which could potentially be improved.

Since WebGL doesn't support threads, there's change that introduces an unthreaded "passthrough" worker. This is the simplest way to get this working, but involves some overhead of creating the messages for the worker in situations where that is not needed. A refactor that uses the worker only on platforms where it's needed should be considred.

refs: RUM-9322