fix: Pin dd-sdk-android dependency to version 2.26.0
Unity needs to target Android SDK v2, without the breaking changes that
are present in the latest snapshot builds. This PR

This PR also commits the changes to the `Datadog Sample` and `Demo Data`
projects that result from opening those projects and building them for
Android in the latest Unity 2022.3 release.