fix: Properly initialize VitalsUpdateFrequency on Android
refs: RUM-12764

Since 1.5.0, `VitalsUpdateFrequency` is configurable via the Unity
project's Datadog Settings, where it defaults to `Average`.

Previously, we only explicitly conveyed that value to the Android SDK
when it was not `None`, which incorrectly resulted in tracking being
enabled in the `None` case, since when no value is explicitly provided,
the Android SDK defaults to `AVERAGE`.

This PR ensures that we convey the configured `VitalsUpdateFrequency`
value to the Android SDK unconditionally.