fix: Restart worker thread if it is detected as dead
If something in the worker thread throws, it can stop that thread preventing Datadog from collecting data from then on. While we're pretty careful about catching possible exceptions in the worker thread, it's possible an exception is slipping through.

I don't have a reproducible case of the worker failing, but I've put in some telemetry to try to catch it, as well as removing some old debug logging from the worker thread.